Boost warning level to 4

+++ apptest/src/base/thread/thread_name.cpp
@@ -27,7 +27,7 @@ void set_debugger_thread_name(const char* name) {
 	THREADNAME_INFO info;
 	info.dwType = 0x1000;
 	info.szName = name;
-	info.dwThreadID = -1;
+	info.dwThreadID = static_cast<DWORD>(-1);
 	info.dwFlags = 0;
 
 #pragma warning(push)
